+++ src/Genius/Resources/AnnotationsResource.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Genius\Resources;
 
-use Genius\Authentication\Scope;
+use Genius\Enum\Scope;
 use stdClass;
 
 /**
@@ -22,7 +22,7 @@ final class AnnotationsResource extends AbstractResource
 
     public function post(array $annotation, array $referent, array $web_page): stdClass
     {
-        $this->requireScope(Scope::SCOPE_CREATE_ANNOTATION);
+        $this->requireScope(Scope::CREATE_ANNOTATION());
 
         return $this->requester->post('annotations', [
             'annotation' => $annotation,
@@ -33,7 +33,7 @@ final class AnnotationsResource extends AbstractResource
 
     public function put(int $id, array $annotation, array $referent, array $web_page): stdClass
     {
-        $this->requireScope(Scope::SCOPE_MANAGE_ANNOTATION);
+        $this->requireScope(Scope::MANAGE_ANNOTATION());
 
         return $this->requester->put(
             sprintf('annotations/%s', $id),
@@ -47,28 +47,28 @@ final class AnnotationsResource extends AbstractResource
 
     public function delete(int $id): stdClass
     {
-        $this->requireScope(Scope::SCOPE_MANAGE_ANNOTATION);
+        $this->requireScope(Scope::MANAGE_ANNOTATION());
 
         return $this->requester->delete(sprintf('annotations/%s', $id));
     }
 
     public function upvote(int $id): stdClass
     {
-        $this->requireScope(Scope::SCOPE_VOTE);
+        $this->requireScope(Scope::VOTE());
 
         return $this->requester->put(sprintf('annotations/%s/upvote', $id));
     }
 
     public function downvote(int $id): stdClass
     {
-        $this->requireScope(Scope::SCOPE_VOTE);
+        $this->requireScope(Scope::VOTE());
 
         return $this->requester->put(sprintf('annotations/%s/downvote', $id));
     }
 
     public function unvote(int $id): stdClass
     {
-        $this->requireScope(Scope::SCOPE_VOTE);
+        $this->requireScope(Scope::VOTE());
 
         return $this->requester->put(sprintf('annotations/%s/unvote', $id));
     }